+++ packages/api-client/src/broadcast/BroadcastAPI.ts
@@ -17,12 +17,12 @@
  *
  */
 
-import {AxiosRequestConfig} from 'axios';
 import {NewOtrMessage as ProtobufOTRMessage} from '@wireapp/protocol-messaging/web/otr';
+import type {AxiosRequestConfig} from 'axios';
 
-import {ClientMismatch, NewOTRMessage} from '../conversation/';
-import {HttpClient} from '../http/';
 import {ValidationError} from '../validation/';
+import type {ClientMismatch, NewOTRMessage} from '../conversation/';
+import type {HttpClient} from '../http/';
 
 export class BroadcastAPI {
   constructor(private readonly client: HttpClient) {}