+++ packages/api-client/src/conversation/OTRRecipients.ts
@@ -17,7 +17,7 @@
  *
  */
 
-import {OTRClientMap} from './';
+import type {OTRClientMap} from './';
 
 export interface OTRRecipients<T extends string | Uint8Array> {
   [userId: string]: OTRClientMap<T>;