+++ packages/core/src/main/conversation/ConversationService.ts
@@ -17,7 +17,7 @@
  *
  */
 
-import {APIClient} from '@wireapp/api-client';
+import type {APIClient} from '@wireapp/api-client';
 import {
   Conversation,
   CONVERSATION_TYPE,
@@ -26,8 +26,8 @@ import {
   NewConversation,
 } from '@wireapp/api-client/src/conversation/';
 import {CONVERSATION_TYPING, ConversationMemberUpdateData} from '@wireapp/api-client/src/conversation/data/';
-import {ConversationMemberLeaveEvent} from '@wireapp/api-client/src/event/';
-import {UserPreKeyBundleMap} from '@wireapp/api-client/src/user/';
+import type {ConversationMemberLeaveEvent} from '@wireapp/api-client/src/event/';
+import type {UserPreKeyBundleMap} from '@wireapp/api-client/src/user/';
 import {
   Asset,
   ButtonAction,
@@ -57,13 +57,13 @@ import {
   PayloadBundleState,
   PayloadBundleType,
 } from '../conversation/';
-import {AssetContent, ClearedContent, DeletedContent, HiddenContent, RemoteData} from '../conversation/content/';
-import {CryptographyService, EncryptedAsset} from '../cryptography/';
+import type {AssetContent, ClearedContent, DeletedContent, HiddenContent, RemoteData} from '../conversation/content/';
+import type {CryptographyService, EncryptedAsset} from '../cryptography/';
 import * as AssetCryptography from '../cryptography/AssetCryptography.node';
 import {MessageBuilder} from './message/MessageBuilder';
 import {MessageService} from './message/MessageService';
 import {MessageToProtoMapper} from './message/MessageToProtoMapper';
-import {
+import type {
   ButtonActionConfirmationMessage,
   ButtonActionMessage,
   CallMessage,