+++ packages/react-ui-kit/src/Form/RoundIconButton.tsx
@@ -18,20 +18,18 @@
  */
 
 /** @jsx jsx */
-import {ObjectInterpolation, jsx} from '@emotion/core';
+import {CSSObject, jsx} from '@emotion/core';
 import React from 'react';
-import {SVGIconProps} from '../Icon/SVGIcon';
+
+import type {SVGIconProps} from '../Icon/SVGIcon';
 import {COLOR} from '../Identity';
-import {Theme} from '../Layout';
+import type {Theme} from '../Layout';
 import {childrenWithDefaultProps} from '../Misc/';
 import {ButtonProps, buttonStyle, filterButtonProps} from './Button';
 
 export interface RoundIconButtonProps<T = HTMLButtonElement> extends ButtonProps<T> {}
 
-export const roundIconButtonStyle: <T>(
-  theme: Theme,
-  props: RoundIconButtonProps<T>,
-) => ObjectInterpolation<undefined> = (theme, props) => ({
+export const roundIconButtonStyle: <T>(theme: Theme, props: RoundIconButtonProps<T>) => CSSObject = (theme, props) => ({
   ...buttonStyle(theme, props),
   alignItems: 'center',
   borderRadius: '50%',