+++ packages/react-ui-kit/src/Form/Select.tsx
@@ -18,9 +18,10 @@
  */
 
 /** @jsx jsx */
-import {ObjectInterpolation, jsx} from '@emotion/core';
+import {CSSObject, jsx} from '@emotion/core';
+
 import {COLOR} from '../Identity';
-import {Theme} from '../Layout';
+import type {Theme} from '../Layout';
 import {filterProps, inlineSVG} from '../util';
 import {InputProps, inputStyle} from './Input';
 
@@ -35,7 +36,7 @@ const ArrowDown = (theme: Theme) => `
   </svg>
 `;
 
-export const selectStyle: <T>(theme: Theme, props: SelectProps<T>) => ObjectInterpolation<undefined> = (
+export const selectStyle: <T>(theme: Theme, props: SelectProps<T>) => CSSObject = (
   theme,
   {disabled = false, markInvalid, ...props},
 ) => ({