+++ packages/react-ui-kit/src/GlobalStyle.tsx
@@ -18,13 +18,14 @@
  */
 
 /** @jsx jsx */
-import {Global, ObjectInterpolation, css, jsx} from '@emotion/core';
+import {Global, CSSObject, css, jsx} from '@emotion/core';
 import emotionNormalize from 'emotion-normalize';
 import {withTheme} from 'emotion-theming';
-import {Theme} from './Layout';
+
+import type {Theme} from './Layout';
 import {textLinkStyle} from './Text/TextLink';
 
-const globalStyles: (theme: Theme) => ObjectInterpolation<undefined> = theme => ({
+const globalStyles: (theme: Theme) => CSSObject = theme => ({
   '*': {
     boxSizing: 'border-box',
   },