+++ packages/react-ui-kit/src/Identity/AvatarGrid.tsx
@@ -18,9 +18,10 @@
  */
 
 /** @jsx jsx */
-import {ObjectInterpolation, jsx} from '@emotion/core';
+import {CSSObject, jsx} from '@emotion/core';
 import React from 'react';
-import {AvatarProps} from '../Identity/';
+
+import type {AvatarProps} from '../Identity/';
 import {IsInViewport} from '../Misc/';
 import {filterProps} from '../util';
 import {Avatar, DEFAULT_AVATAR_SIZE} from './Avatar';
@@ -35,7 +36,7 @@ interface Props<T = HTMLDivElement> extends React.HTMLProps<T> {
   size?: number;
 }
 
-const avatarGridStyle: <T>(props: Props<T>) => ObjectInterpolation<undefined> = ({
+const avatarGridStyle: <T>(props: Props<T>) => CSSObject = ({
   borderWidth,
   size,
   backgroundColor = COLOR.GRAY_DARKEN_48,