+++ packages/react-ui-kit/src/Layout/headerMenu/MenuItems.tsx
@@ -19,19 +19,17 @@
 
 /** @jsx jsx */
 import React from 'react';
-import {ObjectInterpolation, jsx} from '@emotion/core';
+import {CSSObject, jsx} from '@emotion/core';
+
 import {QueryKeys, media} from '../../mediaQueries';
 import {filterProps} from '../../util';
-import {Theme} from '../Theme';
+import type {Theme} from '../Theme';
 
 export interface MenuItemsProps<T = HTMLDivElement> extends React.HTMLProps<T> {
   open?: boolean;
 }
 
-export const menuItemsStyle: <T>(theme: Theme, props: MenuItemsProps<T>) => ObjectInterpolation<undefined> = (
-  theme,
-  props,
-) => ({
+export const menuItemsStyle: <T>(theme: Theme, props: MenuItemsProps<T>) => CSSObject = (theme, props) => ({
   alignSelf: 'center',
   display: 'flex',
   justifySelf: 'end',