+++ packages/react-ui-kit/src/Menu/TabBar.tsx
@@ -18,16 +18,17 @@
  */
 
 /** @jsx jsx */
-import {ObjectInterpolation, jsx} from '@emotion/core';
+import {CSSObject, jsx} from '@emotion/core';
 import React from 'react';
+
 import {COLOR} from '../Identity';
-import {Theme} from '../Layout';
+import type {Theme} from '../Layout';
 import {TextProps, textStyle} from '../Text';
 import {filterProps} from '../util';
 
 export type TabBarProps<T = HTMLDivElement> = React.HTMLProps<T>;
 
-const tabBarStyle: <T>(props: TabBarProps<T>) => ObjectInterpolation<undefined> = ({}) => {
+const tabBarStyle: <T>(props: TabBarProps<T>) => CSSObject = ({}) => {
   return {
     display: 'flex',
     width: '100%',
@@ -46,7 +47,7 @@ export interface TabBarItemProps<T = HTMLSpanElement> extends TextProps<T> {
   active: boolean;
 }
 
-const tabBarItemStyle: <T>(theme: Theme, props: TabBarItemProps<T>) => ObjectInterpolation<undefined> = (
+const tabBarItemStyle: <T>(theme: Theme, props: TabBarItemProps<T>) => CSSObject = (
   theme,
   {
     block = true,